refactor(WinningMoney.java) : 당첨금액ENUM명 변경 및 테스트 시그니처 메서드 작성
