+++ docs/README.md
@@ -114,6 +114,13 @@
 - [x] 일치 갯수를 판별.
     - [x] 로또 번호와 당첨 번호를 비교하여 일치하는 갯수를 반환한다.
 
+### 당첨 금액
+
+- [ ] 일치 갯수가 3개이면 5_000원을 수령한다.
+- [ ] 일치 갯수가 4개이면 50_000원을 수령한다.
+- [ ] 일치 갯수가 5개이면 1_500_000원을 수령한다.
+- [ ] 일치 갯수가 6개이면 2_000_000_000원을 수령한다.
+
 #### 통계
 
 - [x] 수익률을 계산.
