+++ src/test/java/lotto/domain/WinningMoneyTest.java
@@ -1,0 +1,32 @@
+package lotto.domain;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+@DisplayName("당첨금액")
+class WinningMoneyTest {
+
+    @DisplayName("일치 갯수가 3개이면 5_000원을 수령한다.")
+    @Test
+    void three_match() {
+
+    }
+
+    @DisplayName("일치 갯수가 4개이면 50_000원을 수령한다.")
+    @Test
+    void four_match() {
+
+    }
+
+    @DisplayName("일치 갯수가 5개이면 1_500_000원을 수령한다.")
+    @Test
+    void five_match() {
+
+    }
+
+    @DisplayName("일치 갯수가 6개이면 2_000_000_000원을 수령한다.")
+    @Test
+    void six_match() {
+
+    }
+}
